android: Allow clear-text traffic for now
We need to be able to talk to https servers with an official lets-encrypt
signed certificate as well as local servers that cannot have an official
certificate, so SSl pinning as the nativescript-https does isn't going to work.
I'm also not aware of a way of installing a self-signed certificate to allow
traffic for.